Refactor Capabilities command
