+++ src/commands/mod.rs
@@ -1,7 +1,7 @@
 // Licensed under the Apache-2.0 license
 
 pub mod algorithms_rsp;
-pub mod capabilities_rsp;
+pub mod capabilities;
 pub mod certificate_rsp;
 pub mod challenge_auth_rsp;
 pub mod chunk_get_rsp;
